Implemented crud commands on musicbase
 - huh, tons of classes created
 - consider another repackage, a big chunk of new classes ariven
 - but the musicbase create-rename-move-remove commands are done

+++ jmop-src/jmop-common-storages/src/main/java/cz/martlin/jmop/common/storages/bundlesdir/BundlesDirsStorage.java
@@ -196,7 +196,10 @@ public class BundlesDirsStorage implements BaseMusicbaseStorage {
 	@Override
 	public void removeTrack(Track track) throws JMOPMusicbaseException {
 		File oldTrackFile = locator.trackFile(track);
-		fs.deleteFile(oldTrackFile);
+		
+		if (oldTrackFile.exists()) {
+			fs.deleteFile(oldTrackFile);
+		}
 	}
 
 	@Override
